chore: autofix corruption (rehydrate/re-quarantine/syntax repair) batch offset=0
